docker_compose/django/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
cfgv==3.3.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.6
django-crispy-forms==1.14.0
django-model-utils==4.2.0
djangorestframework==3.13.1
filelock==3.7.1
identify==2.5.1
nodeenv==1.7.0
packaging==21.3
platformdirs==2.5.2
pre-commit==2.20.0
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.15.1
wrapt==1.14.1
